refactor(RoundResult): 승패판단 로직 분리
